requirements-dev.txt
numpy==1.26.4
pandas==2.2.0
pandas-stubs~=2.2.0
python-dotenv==1.0.1
pyarrow==15.0.0
openpyxl==3.1.2
coverage~=7.1.0
flake8~=6.0.0
docstr-coverage~=2.2.0
sphinx~=6.1.3
sphinx-rtd-theme~=1.2.0
sphinx-rtd-dark-mode~=1.2.4
sphinx_mdinclude~=0.5.3
mypy~=1.3.0
testbook~=0.4.2
pytest~=8.1.1
jupyter~=1.0.0
biopython~=1.83
dnachisel~=3.2.11
types-requests~=2.31
xlrd~=2.0.1
static-frame~=2.13.0